requirements.txt
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy
google-genai
python-dotenv